+++ autoapi/mappers/go.py
@@ -30,6 +30,8 @@ class GoSphinxMapper(SphinxMapperBase):
             if data:
                 self.paths[_dir] = data
 
+        return True
+
     def read_file(self, path, **kwargs):
         """Read file input into memory, returning deserialized objects